chore: simplify code
